requirements.txt
faker-clickstream
dbt-core
dbt-postgres
psycopg2-binary 
python-dotenv
pytest